fix(Create User): Allow the system administrator to create new user accounts
Modified the CreateUserDto to take the input fields from the modal components. Integrated the createUser function in useStore.ts. Added validation for password confirmation

Currently the useStore shows the createUser function to send a request to the backend. Removed the name column from the user entity. Removed the name column from CreateUserDto and RegisterDto. Changed backend address to localhost:3001 in useStore