fix(CreateExerciseComponent.tsx): Fix the memoryId field returning null
The memoryId field returned null because the data.memoryId became an object with two parts: label and value. The number sent to the database is the value property in memoryId: parseInt(data.memoryId.value). This number is an int equal to the label divided by 64, so selecting 64 would return 1, 128 would return 2, and so on...