fix(MarkdownTailwind): Fix the styles of the markdown content
Hidden the KaTeX duplicates with the use of a global style. Capped the max size of the news article's cover picture to 1/3 of the card. Added new rules of typography and prose to fix the color setting in dark mode. Capped the card's size to 11/12 of the page's width on large screens. Centered the images in the markdown body with m-auto.