fix(fixTickets): Display both versions of a modified item
Adjusted the Ticket page to display both versions of a modified item. Rewritten the whole process of updating news and exercises, as they replaced the original version of the item. Fixed a query that did not return the body of an exercise in the ticket page. Removed console statements. Modified the creation of exercises to accept null values for optional fields instead of throwing errors. Fixed a bug in the CreateExcerciseComponent that set the id of a memory limit as undefined. Moved the header of the exercise into a new client component to dynamically hide the ReportErrorButton according to the current route. Removed unused imports.