fix(TicketPage): Display the Exercise Tickets
As of currently, the Exercise tickets have a problem displaying due to a difference between the server rendered version and the client rendered version of the page. Which components should be client side and server side? I'm not sure

WIP, the ticket page for an exercise still throws a "TypeError: Cannot read properties of undefined (reading 'default')"