fix(Reports): Let the trainers and admins close reports
Modified the DisplayReportComponent to provide a link to the item to which the report is asociated. Created a new endpoint to close reports. Modified the profile table to only display open reports. Modified the ReportCardComponent to only receive normal text instead of MDX. Modified the ReportCardComponent to display error messages. Added the LoggerService to the report Controller.