fix(User CRUD): Fix problems with the user operations
Changed the text for the CardWithHeaderComponent as requested. Modified the services to keep track of the time the item was created and who created the item. Modified the margin in the buttons of the TabComponent. Fixed the unexpected selection of a time limit in the CreateExerciseComponent. Changed the CSS selector for the double katex in the globals.css