fix(cloudStorage): Fix ENOENT error when uploading image
Part of the commented code was important to upload images to the Cloud Storage bucket. The commented code created a local copy of the image for upload, which is a necessary step for the flow of data during the creation of a news article. The commented code also had an incorrect call of an env variable, attempting to use the name of the bucket as part of the path of the local copy of the image to upload, this has been fixed.

MAKE SURE TO HAVE A "publicAssets" FOLDER